ASoC: fix [XJ3-4097] Fix compile warning about type conversion no_3rd_skip
Summary:
fix compile warning about type conversion

Test Plan: no

Reviewers: zhen01.zhang, yunqian.wang

Reviewed By: zhen01.zhang, yunqian.wang

Differential Revision: https://cr.hobot.cc/D161922

+++ sound/soc/codecs/ac10x.h
@@ -80,7 +80,7 @@ struct ac10x_priv {
 
 	#ifdef CONFIG_AC101_SWITCH_DETECT
 	struct gpio_desc* gpiod_irq;
-	long irq;
+	unsigned int irq;
 	volatile int irq_cntr;
 	volatile int pullout_cntr;
 	volatile int state;
